Refactor receipt printing into reusable template
